+++ AirCompany.Domain/Model/AircraftModel.cs
@@ -1,4 +1,4 @@
-﻿namespace AirCompany.Domain.Models;
+﻿namespace AirCompany.Domain.Model;
 
 /// <summary>
 /// Represents a specific <see cref="AircraftModel"/> (e.g., "A320neo").
