+++ AirCompany.Domain/Model/Flight.cs
@@ -1,4 +1,4 @@
-﻿namespace AirCompany.Domain.Models;
+﻿namespace AirCompany.Domain.Model;
 
 
 /// <summary>
