+++ AirCompany.Domain/Model/Ticket.cs
@@ -1,8 +1,8 @@
-﻿namespace AirCompany.Domain.Models;
+﻿namespace AirCompany.Domain.Model;
 
 /// <summary>
-/// Represents a <see cref="Ticket"/> issued to a <see cref="Models.Passenger"/> 
-/// for a specific <see cref="Models.Flight"/>.
+/// Represents a <see cref="Ticket"/> issued to a <see cref="Model.Passenger"/> 
+/// for a specific <see cref="Model.Flight"/>.
 /// </summary>
 public class Ticket
 {
@@ -12,12 +12,12 @@ public class Ticket
     public required int Id { get; set; }
 
     /// <summary>
-    /// The <see cref="Models.Flight"/> associated with this <see cref="Ticket"/>.
+    /// The <see cref="Model.Flight"/> associated with this <see cref="Ticket"/>.
     /// </summary>
     public required Flight Flight { get; set; }
 
     /// <summary>
-    /// The <see cref="Models.Passenger"/> who owns this <see cref="Ticket"/>.
+    /// The <see cref="Model.Passenger"/> who owns this <see cref="Ticket"/>.
     /// </summary>
     public required Passenger Passenger { get; set; }
 
@@ -27,7 +27,7 @@ public class Ticket
     public required string SeatNumber { get; set; }
 
     /// <summary>
-    /// Indicates whether the <see cref="Models.Passenger"/> has hand luggage for this <see cref="Ticket"/>.
+    /// Indicates whether the <see cref="Model.Passenger"/> has hand luggage for this <see cref="Ticket"/>.
     /// </summary>
     public bool? HasHandLuggage { get; set; }
